requirements.txt
beautifulsoup4
python-nmap
requests
urllib3